refactor: move cdf_aggregate
